+++ src/pygeoprocessing/multiprocessing/__init__.py
@@ -1,1 +1,2 @@
 from .raster_calculator import raster_calculator
+__all__ = ['raster_calculator']