+++ HISTORY.rst
@@ -3,8 +3,19 @@ Release History
 
 Unreleased Changes
 ------------------
+* Changed parameters in ``convolve_2d`` to allow API to set
+  ``ignore_nodata_and_edges=True`` while ``mask_nodata=False`` and updated
+  docstring to indicate this is useful in cases such as filling nodata holes
+  in missing datasets. Additionally added a logger ``debug`` message to note
+  this "unusual" setting of these parameters in case of accidental usage
+  which could be noted during development.
+
+2.1.2 (2020-12-03)
+------------------
 * ``pygeoprocessing.warp_raster`` now raises a ``ValueError`` when an invalid
   resampling method is provided.
+* Fixed bug in ``pygeoprocessing.warp_raster`` that would not properly handle
+  GDAL Byte type signing when masking warped raster with a vector.
 * Fixed issue in ``convolve_2d`` that would cause excessive memory use
   leading to out of memory errors.
 * Fixed issue in ``convolve_2d`` that could lead to a file removal race